Refactor + remove centering
It was fun to print to the center, but it's code that's not really going
to be useful, so let's get rid of it.

At the same time, the functions should know their own address, so let's
stop passing the vga location around all the time. Main looks so much
nicer now!