Fix hardcoded addresses so that the kernel can be linked to 1M again
The `offset_middle` field is hardcoded and needs to be adjusted to 1 instead of 2. Similarly, we need to subtract `0x100000` instead of `0x200000`.

We still need to add the sections to the `.boot` output section so that `offset_middle` stays 1, even if our kernel grows significantly. The rest of the linker script is reverted to the previous version.