Users no longer need their own libcore
This commit, along with the creation of the intermezzos/libcore repo,
means that developers of intermezzOS need to do significantly less work
to get things going. Here's how it works:

We support a specific nightly. In this case. 2016/04/12. Manual work
will be need to be done to move this forward, whenever we decide to.
A bit less flexible, but frankly, pinning to a specific nightly is
better; it's more stable that way.

The intermezzos/libcore repo contains a specific nightly's libcore with
the floating-point support removed.

The build system will now clone that repository down, and build it. One
possible weakness: this just downloads HEAD. What will happen when
we upgrade? Will existing users get sudden errors? I'll file an issue
about that, and mention this commit, so github links them.

We then pass a path to the final libcore.rlib to cargo when building
the Rust code in the kernel. Hooray!

:confetti_ball: :shipit:

Fixes @intermezzOS/book#26