Simplify boot.asm a bit
So, here's the thing: the code we had before is a bit nicer in that it
uses functions, but it's also a bit more complex in that we had to set
up a stack, etc. So I'm going to make the code "worse" in one sense,
but eliminating a concept or two. It's fine.