Pin libcore to a certain revision
Initially a concern in b294480f404b1fccbf745ea53affca9b48b9e482, this
commit now checks out a specific hash of our libcore, so that when we
update it, people's stuff won't break, until we update it here as well.
This allows us to experiment a bit more freely with upgrades, but not
make them 'official' until the SHA here is updated.

Fixes #25