Move more asm into Rust
This commit does two things: first, it removes the setup_SSE, as we are
not actually using SSE instructions. This was introduced when following
Phil's tutorial long ago, but now we have a libcore that doesn't use SSE
at all, so this should be fine.

Second, we move the PIC stuff into Rust. It's just some outb/inb calls,
no need to keep it in assembly.