Replace old vga driver with the new one.
This technically leads to a functional regression, but that's okay for
now. The new vga driver does not have the ability to blink the cursor,
that's being tracked over at https://github.com/intermezzOS/vga/issues/3

This also needs https://github.com/intermezzOS/vga/pull/2 to be merged
before it will work here.