Use xargo to build
This is much much nicer than maintaining a port of libcore. Thanks
@japaric! :heart:

This also means that the "hello cargo" chapter of the book will need to
be redone.

I also re-organized the .json file to be in alphabetical order and fixed
some indentation issues. I also added +soft-float, which is needed since
we no longer have the no-float patch.

I also had to update the x86 dependency to work with current nightly.

There were some updates to the travis.yml: removing the specific VM
version of Trusty, which meant installing grub-common.

This also means that we are no longer tied to a specific Rust version.

Finally, --gc-sections was needed. We used to have a no-float patch for
libcore when we compiled it custom, but xargo doesn't know how to do
that. --gc-sections will get rid of the unused calls instead.