Remove unsafe from Vga's constructor.
Turns out, we don't need to worry about this! We can accept a mutable
slice instead, and panic if it's the wrong length.

This now means that the VGA driver is 100% safe, and the only mention of
unsafety in the kernel at all is the creation of the slice to video
memory.

Since I was changing the signature of Vga::new, I also cleaned up some
of the const fn stuff that used to matter when we had a static Vga, but
now do not. :metal: