kernel-context -> libintermezzos
I originally pulled the context out into its own crate so that the other
ones could depend on it. When it was part of the kernel binary, that
wasn't possible. But it doesn't make sense as its own crate; it's
actually part of the intermezzos crate.

So, this commit introduces a new library intermezzos crate, which the
binary then depends upon. Other crates will be able to include it as
well, in the future.