Remove example source code.
The example source code is a good thing, but it doesn't really fit in
this repository. As such, remove all of it and move the old src
directory into the top-level.

For now, I've saved the example source code as tags:
https://github.com/intermezzOS/kernel/tags

I might move it into the book repository, since it is, after all, for
the book. We'll see. :smile: