Super-rudimentary state machine for the keyboard driver
The PS/2 port sends pressed and released codes corresponding to
the user pressing and releasing keys. To accurately turn those into
printable and non-printable key codes we need a state machine to
model the state of the keyboard. If we don't keep track of state
then we cannot know when/if modifier keys (eg. shift, alt, etc.)
should be applied to a given key-stroke.